add settings to plugins
