+++ packages/graph/src/plugins/drag/settings.ts
@@ -1,0 +1,11 @@
+export type DraggableGraphSettings = {
+  /**
+   * whether the nodes on the graph are draggable
+   * @default true
+   */
+  draggable: boolean;
+};
+
+export const DEFAULT_DRAGGABLE_SETTINGS: DraggableGraphSettings = {
+  draggable: true,
+};
